added safely with use of crontab
since the code runs in an interval of 15 minutes and the previous changes looped the code till an internet connection was established before ending, i added a safety feature which automatically ends the code after a couple of tried. this will allow the code to end and crontab will run the code again without crossing over.